Moved tests to their own folder and simplified.
