--- tests/MongoDB.Driver.Tests/MultipleRegistriesTests.cs
@@ -1,1127 +1,0 @@
-/* Copyright 2010-present MongoDB Inc.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-using System;
-using System.Linq;
-using FluentAssertions;
-using MongoDB.Bson;
-using MongoDB.Bson.Serialization;
-using MongoDB.Bson.Serialization.Attributes;
-using MongoDB.Bson.Serialization.Conventions;
-using MongoDB.Bson.Serialization.Serializers;
-using MongoDB.Driver.Core.TestHelpers.XunitExtensions;
-using MongoDB.Driver.Linq;
-using MongoDB.Driver.Tests.Linq.Linq3Implementation;
-using Xunit;
-
-namespace MongoDB.Driver.Tests
-{
-    [Trait("Category", "Integration")]
-    public class MultipleRegistriesTests
-    {
-        private readonly string _defaultObjectIdString = "6797b56bf5495bf53aa3078f";
-        private readonly ObjectId _defaultId = ObjectId.Parse("6797b56bf5495bf53aa3078f");
-
-        [Fact]
-        public void TestSerialization()
-        {
-            RequireServer.Check();
-
-            // The first section demonstrates that the class maps are also separated
-            {
-                var client = CreateClient();
-                var collection = GetTypedCollection<Person>(client);
-                var bsonCollection = GetUntypedCollection(client);
-
-                var person = new Person { Id = _defaultId, Name = "Mario", Age = 24 };
-                collection.InsertOne(person);
-
-                var retrieved = bsonCollection.FindSync(FilterDefinition<BsonDocument>.Empty).ToList().Single();
-                var toString = retrieved.ToString();
-
-                var expectedVal =
-                    $$"""{ "_id" : { "$oid" : "{{_defaultObjectIdString}}" }, "Name" : "Mario", "Age" : 24 }""";
-                Assert.Equal(expectedVal, toString);
-            }
-
-            {
-                var customDomain = BsonSerializationDomain.CreateWithDefaultConfiguration("Test");
-                customDomain.RegisterSerializer(new CustomStringSerializer("test1"));
-
-                var client = CreateClientWithDomain(customDomain);
-                var collection = GetTypedCollection<Person>(client);
-                var bsonCollection = GetUntypedCollection(client);
-
-                var person = new Person { Id = _defaultId, Name = "Mario", Age = 24 };
-                collection.InsertOne(person);
-
-                var retrievedAsBson = bsonCollection.FindSync(FilterDefinition<BsonDocument>.Empty).ToList().Single();
-                var toString = retrievedAsBson.ToString();
-
-                var expectedVal =
-                    $$"""{ "_id" : { "$oid" : "{{_defaultObjectIdString}}" }, "Name" : "Mariotest1", "Age" : 24 }""";
-                Assert.Equal(expectedVal, toString);
-
-                var retrievedTyped = collection.FindSync(FilterDefinition<Person>.Empty).ToList().Single();
-                Assert.Equal("Mario", retrievedTyped.Name);
-            }
-
-            {
-                var customDomain = BsonSerializationDomain.CreateWithDefaultConfiguration("Test");
-                customDomain.RegisterSerializer(new CustomStringSerializer("test2"));
-
-                var client = CreateClientWithDomain(customDomain);
-                var collection = GetTypedCollection<Person>(client);
-                var bsonCollection = GetUntypedCollection(client);
-
-                var person = new Person { Id = _defaultId, Name = "Mario", Age = 24 };
-                collection.InsertOne(person);
-
-                var retrievedAsBson = bsonCollection.FindSync(FilterDefinition<BsonDocument>.Empty).ToList().Single();
-                var toString = retrievedAsBson.ToString();
-
-                var expectedVal =
-                    $$"""{ "_id" : { "$oid" : "{{_defaultObjectIdString}}" }, "Name" : "Mariotest2", "Age" : 24 }""";
-                Assert.Equal(expectedVal, toString);
-
-                var retrievedTyped = collection.FindSync(FilterDefinition<Person>.Empty).ToList().Single();
-                Assert.Equal("Mario", retrievedTyped.Name);
-            }
-        }
-
-        [Fact]
-        public void TestMultipleDomainSimultaneously()
-        {
-            RequireServer.Check();
-
-            var objectId1 = ObjectId.GenerateNewId();
-            var objectId2 = ObjectId.GenerateNewId();
-
-            var client = CreateClient();
-            var collection = GetTypedCollection<Person>(client);
-            var bsonCollection = GetUntypedCollection(client);
-
-            var customDomain = BsonSerializationDomain.CreateWithDefaultConfiguration("Test");
-            customDomain.RegisterSerializer(new CustomStringSerializer("test1"));
-            var client2 = CreateClientWithDomain(customDomain);
-            var collection2 = GetTypedCollection<Person>(client2);
-            var bsonCollection2 = GetUntypedCollection(client2);
-
-            var person = new Person { Id = objectId1, Name = "Mario", Age = 24 };
-            var person2 = new Person { Id = objectId2, Name = "Mario", Age = 24 };
-            collection.InsertOne(person);
-            collection2.InsertOne(person2);
-
-            var retrieved = bsonCollection.FindSync(Builders<BsonDocument>.Filter.Eq("_id", objectId1)).ToList().Single();
-            var expectedVal =
-                $$"""{ "_id" : { "$oid" : "{{objectId1.ToString()}}" }, "Name" : "Mario", "Age" : 24 }""";
-            Assert.Equal(expectedVal, retrieved.ToString());
-
-            var retrievedAsBson = bsonCollection2.FindSync(Builders<BsonDocument>.Filter.Eq("_id", objectId2)).ToList().Single();
-            var expectedVal2 =
-                $$"""{ "_id" : { "$oid" : "{{objectId2.ToString()}}" }, "Name" : "Mariotest1", "Age" : 24 }""";
-            Assert.Equal(expectedVal2, retrievedAsBson.ToString());
-
-            var retrievedTyped = collection2.FindSync(p => p.Id == objectId2).ToList().Single();
-            Assert.Equal("Mario", retrievedTyped.Name);
-        }
-
-        [Fact]
-        public void TestDeserialization()
-        {
-            RequireServer.Check();
-
-            {
-                var client = CreateClient();
-                var collection = GetTypedCollection<Person1>(client);
-
-                var person = new Person1 { Id = _defaultId, Name = "Mariotest", Age = 24 };
-                collection.InsertOne(person);
-            }
-
-            {
-                var customDomain = BsonSerializationDomain.CreateWithDefaultConfiguration("Test");
-                customDomain.RegisterSerializer(new CustomStringSerializer());
-
-                var client = CreateClientWithDomain(customDomain, dropCollection: false);
-                var collection = GetTypedCollection<Person>(client);
-
-                var retrievedTyped = collection.FindSync(FilterDefinition<Person>.Empty).ToList().Single();
-                Assert.Equal("Mario", retrievedTyped.Name);
-            }
-        }
-
-        [Fact]
-        public void TestLinq()
-        {
-            RequireServer.Check();
-
-            var customDomain = BsonSerializationDomain.CreateWithDefaultConfiguration("Test");
-            customDomain.RegisterSerializer(new CustomStringSerializer());
-
-            var client = CreateClientWithDomain(customDomain);
-            var collection = GetTypedCollection<Person>(client);
-            var untypedCollection = GetUntypedCollection(client);
-
-            var person = new Person { Id = _defaultId, Name = "Mario", Age = 24 };
-            collection.InsertOne(person);
-
-            var retrievedAsBson = untypedCollection.FindSync(FilterDefinition<BsonDocument>.Empty).ToList().Single();
-            var toString = retrievedAsBson.ToString();
-
-            var expectedVal =
-                $$"""{ "_id" : { "$oid" : "{{_defaultObjectIdString}}" }, "Name" : "Mariotest", "Age" : 24 }""";
-            Assert.Equal(expectedVal, toString);
-
-            var retrievedTyped = collection.AsQueryable().Where(x => x.Name == "Mario").ToList();  //The string serializer is correctly serializing "Mario" to "Mariotest"
-            Assert.NotEmpty(retrievedTyped);
-        }
-
-        [Fact]
-        public void TestLinqConcatTranslatesSecondQueryableUnderOuterDomain()
-        {
-            RequireServer.Check();
-
-            // Outer domain uses default naming; inner domain uses lowercase field names for Person.
-            var outerDomain = BsonSerializationDomain.CreateWithDefaultConfiguration("outer");
-            var innerDomain = BsonSerializationDomain.CreateWithDefaultConfiguration("inner");
-            var lowercasePack = new ConventionPack(innerDomain);
-            lowercasePack.AddMemberMapConvention(
-                "LowerCaseElementName",
-                m => m.SetElementName(m.MemberName.ToLower()));
-            innerDomain.ConventionRegistry.Register("lower", lowercasePack, t => t == typeof(Person));
-
-            var outerClient = DriverTestConfiguration.CreateMongoClient(c => c.SerializationDomain = outerDomain);
-            var innerClient = DriverTestConfiguration.CreateMongoClient(c => c.SerializationDomain = innerDomain);
-            var databaseName = DriverTestConfiguration.DatabaseNamespace.DatabaseName;
-            var outerColl = outerClient.GetDatabase(databaseName).GetCollection<Person>("concat_outer");
-            var innerColl = innerClient.GetDatabase(databaseName).GetCollection<Person>("concat_inner");
-
-            var queryable = outerColl.AsQueryable().Concat(innerColl.AsQueryable().Where(p => p.Name == "Alice"));
-
-            var stages = Linq3TestHelpers.Translate<Person, Person>(queryable);
-            var unionWith = stages.Single(s => s.Contains("$unionWith"));
-            // Outer-domain translation means the inner $match references "Name", not the inner domain's lowercase "name".
-            unionWith.ToJson().Should().Contain("\"Name\" : \"Alice\"");
-            unionWith.ToJson().Should().NotContain("\"name\" : \"Alice\"");
-        }
-
-        [Fact]
-        public void TestConventions()
-        {
-            RequireServer.Check();
-
-            var customDomain = BsonSerializationDomain.CreateWithDefaultConfiguration("Test");
-
-            // Register an id generator convention that uses a custom ObjectIdGenerator
-            customDomain.RegisterIdGenerator(typeof(ObjectId), new CustomObjectIdGenerator());
-
-            //Register a convention to use lowercase for all fields on the Person class
-            var pack = new ConventionPack(customDomain);
-            pack.AddMemberMapConvention(
-                "LowerCaseElementName",
-                m => m.SetElementName(m.MemberName.ToLower()));
-            customDomain.ConventionRegistry.Register("myPack", pack, t => t == typeof(Person));
-
-            var client = CreateClientWithDomain(customDomain);
-            var collection = GetTypedCollection<Person>(client);
-            var untypedCollection = GetUntypedCollection(client);
-
-            var person = new Person { Name = "Mario", Age = 24 };  //Id is not set, so the custom ObjectIdGenerator should be used
-            collection.InsertOne(person);
-
-            var retrievedAsBson = untypedCollection.FindSync(FilterDefinition<BsonDocument>.Empty).ToList().Single();
-            var toString = retrievedAsBson.ToString();
-
-            var expectedVal =
-                $$"""{ "_id" : { "$oid" : "{{_defaultObjectIdString}}" }, "name" : "Mario", "age" : 24 }""";
-            Assert.Equal(expectedVal, toString);
-        }
-
-        [Fact]
-        public void TestDiscriminators()
-        {
-            RequireServer.Check();
-
-            var customDomain = BsonSerializationDomain.CreateWithDefaultConfiguration("Test");
-
-            customDomain.ClassMapRegistry.RegisterClassMap<BasePerson>(cm =>
-            {
-                cm.AutoMap();
-                cm.SetDiscriminator("bp");
-                cm.SetIsRootClass(true);
-            });
-
-            customDomain.ClassMapRegistry.RegisterClassMap<DerivedPerson1>(cm =>
-            {
-                cm.AutoMap();
-                cm.SetDiscriminator("dp1");
-                cm.MapMember( m => m.ExtraField1).SetSerializer(new CustomStringSerializer());
-            });
-
-            customDomain.ClassMapRegistry.RegisterClassMap<DerivedPerson2>(cm =>
-            {
-                cm.AutoMap();
-                cm.SetDiscriminator("dp2");
-                cm.MapMember( m => m.ExtraField2).SetSerializer(new CustomStringSerializer());
-            });
-
-            var client = CreateClientWithDomain(customDomain);
-            var collection = GetTypedCollection<BasePerson>(client);
-            var untypedCollection = GetUntypedCollection(client);
-
-            var bp1 = new DerivedPerson1 { Name = "Alice", Age = 30, ExtraField1 = "Field1" };
-            var bp2 = new DerivedPerson2 { Name = "Bob", Age = 40, ExtraField2 = "Field2" };
-            collection.InsertMany([bp1, bp2]);
-
-            var retrieved1 = untypedCollection.FindSync(Builders<BsonDocument>.Filter.Eq("_id", bp1.Id)).ToList().Single();
-            var retrieved2 = untypedCollection.FindSync(Builders<BsonDocument>.Filter.Eq("_id", bp2.Id)).ToList().Single();
-
-            var expectedDiscriminator1 =
-                $"""_t" : ["bp", "dp1"]""";
-            var expectedDiscriminator2 =
-                $"""_t" : ["bp", "dp2"]""";
-            Assert.Contains(expectedDiscriminator1, retrieved1.ToString());
-            Assert.Contains(expectedDiscriminator2, retrieved2.ToString());
-
-            //Aggregate with OfType
-            var retrievedDerivedPerson1 = collection.Aggregate().OfType<DerivedPerson1>().Single();
-            var retrievedDerivedPerson2 = collection.Aggregate().OfType<DerivedPerson2>().Single();
-
-            AssertDerivedPerson1(bp1, retrievedDerivedPerson1);
-            AssertDerivedPerson2(bp2, retrievedDerivedPerson2);
-
-            //AppendStage with OfType
-            retrievedDerivedPerson1 = collection.AsQueryable().AppendStage(PipelineStageDefinitionBuilder.OfType<BasePerson, DerivedPerson1>()).Single();
-            retrievedDerivedPerson2 = collection.AsQueryable().AppendStage(PipelineStageDefinitionBuilder.OfType<BasePerson, DerivedPerson2>()).Single();
-
-            AssertDerivedPerson1(bp1, retrievedDerivedPerson1);
-            AssertDerivedPerson2(bp2, retrievedDerivedPerson2);
-
-            //LINQ with OfType
-            retrievedDerivedPerson1 = collection.AsQueryable().OfType<DerivedPerson1>().Single();
-            retrievedDerivedPerson2 = collection.AsQueryable().OfType<DerivedPerson2>().Single();
-
-            AssertDerivedPerson1(bp1, retrievedDerivedPerson1);
-            AssertDerivedPerson2(bp2, retrievedDerivedPerson2);
-
-            //Facet with OfType
-
-            var pipeline1 = PipelineDefinition<BasePerson, DerivedPerson1>.Create( new [] {
-                PipelineStageDefinitionBuilder.OfType<BasePerson, DerivedPerson1>() });
-            var facet1 = AggregateFacet.Create("facet1", pipeline1);
-
-            var pipeline2 = PipelineDefinition<BasePerson, DerivedPerson2>.Create( new [] {
-                PipelineStageDefinitionBuilder.OfType<BasePerson, DerivedPerson2>() });
-            var facet2 = AggregateFacet.Create("facet2", pipeline2);
-
-            var result = collection.Aggregate().Facet(facet1, facet2).Single().Facets;
-            retrievedDerivedPerson1 = result[0].Output<DerivedPerson1>().Single();
-            retrievedDerivedPerson2 = result[1].Output<DerivedPerson2>().Single();
-
-            AssertDerivedPerson1(bp1, retrievedDerivedPerson1);
-            AssertDerivedPerson2(bp2, retrievedDerivedPerson2);
-
-            //Find with OfType
-            var retrievedBasePerson1 = collection.FindSync(Builders<BasePerson>.Filter.OfType<DerivedPerson1>()).Single();
-            var retrievedBasePerson2 = collection.FindSync(Builders<BasePerson>.Filter.OfType<DerivedPerson2>()).Single();
-
-            AssertBasePerson(bp1, retrievedBasePerson1);
-            AssertBasePerson(bp2, retrievedBasePerson2);
-
-            void AssertDerivedPerson1(DerivedPerson1 expected, DerivedPerson1 retrieved)
-            {
-                AssertBasePerson(expected, retrieved);
-                Assert.Equal(expected.ExtraField1, retrieved.ExtraField1);
-            }
-
-            void AssertDerivedPerson2(DerivedPerson2 expected, DerivedPerson2 retrieved)
-            {
-                AssertBasePerson(expected, retrieved);
-                Assert.Equal(expected.ExtraField2, retrieved.ExtraField2);
-            }
-
-            void AssertBasePerson(BasePerson expected, BasePerson retrieved)
-            {
-                Assert.Equal(expected.Id, retrieved.Id);
-                Assert.Equal(expected.Name, retrieved.Name);
-                Assert.Equal(expected.Age, retrieved.Age);
-            }
-        }
-
-        [Fact]
-        public void TestDiscriminatorsWithAttributes()
-        {
-            RequireServer.Check();
-
-            var customDomain = BsonSerializationDomain.CreateWithDefaultConfiguration("Test");
-            customDomain.RegisterSerializer(new CustomStringSerializer());
-
-            var client = CreateClientWithDomain(customDomain);
-            var collection = GetTypedCollection<BasePersonAttribute>(client);
-            var untypedCollection = GetUntypedCollection(client);
-
-            var bp1 = new DerivedPersonAttribute1 { Name = "Alice", Age = 30, ExtraField1 = "Field1" };
-            var bp2 = new DerivedPersonAttribute2 { Name = "Bob", Age = 40, ExtraField2 = "Field2" };
-            collection.InsertMany([bp1, bp2]);
-
-            var retrieved1 = untypedCollection.FindSync(Builders<BsonDocument>.Filter.Eq("_id", bp1.Id)).ToList().Single();
-            var retrieved2 = untypedCollection.FindSync(Builders<BsonDocument>.Filter.Eq("_id", bp2.Id)).ToList().Single();
-
-            var expectedDiscriminator1 =
-                $"""_t" : ["bp", "dp1"]""";
-            var expectedDiscriminator2 =
-                $"""_t" : ["bp", "dp2"]""";
-            Assert.Contains(expectedDiscriminator1, retrieved1.ToString());
-            Assert.Contains(expectedDiscriminator2, retrieved2.ToString());
-
-            //Aggregate with OfType
-            var retrievedDerivedPerson1 = collection.Aggregate().OfType<DerivedPersonAttribute1>().Single();
-            var retrievedDerivedPerson2 = collection.Aggregate().OfType<DerivedPersonAttribute2>().Single();
-
-            AssertDerivedPerson1(bp1, retrievedDerivedPerson1);
-            AssertDerivedPerson2(bp2, retrievedDerivedPerson2);
-
-            //AppendStage with OfType
-            retrievedDerivedPerson1 = collection.AsQueryable().AppendStage(PipelineStageDefinitionBuilder.OfType<BasePersonAttribute, DerivedPersonAttribute1>()).Single();
-            retrievedDerivedPerson2 = collection.AsQueryable().AppendStage(PipelineStageDefinitionBuilder.OfType<BasePersonAttribute, DerivedPersonAttribute2>()).Single();
-
-            AssertDerivedPerson1(bp1, retrievedDerivedPerson1);
-            AssertDerivedPerson2(bp2, retrievedDerivedPerson2);
-
-            //LINQ with OfType
-            retrievedDerivedPerson1 = collection.AsQueryable().OfType<DerivedPersonAttribute1>().Single();
-            retrievedDerivedPerson2 = collection.AsQueryable().OfType<DerivedPersonAttribute2>().Single();
-
-            AssertDerivedPerson1(bp1, retrievedDerivedPerson1);
-            AssertDerivedPerson2(bp2, retrievedDerivedPerson2);
-
-
-            void AssertDerivedPerson1(DerivedPersonAttribute1 expected, DerivedPersonAttribute1 retrieved)
-            {
-                AssertBasePerson(expected, retrieved);
-                Assert.Equal(expected.ExtraField1, retrieved.ExtraField1);
-            }
-
-            void AssertDerivedPerson2(DerivedPersonAttribute2 expected, DerivedPersonAttribute2 retrieved)
-            {
-                AssertBasePerson(expected, retrieved);
-                Assert.Equal(expected.ExtraField2, retrieved.ExtraField2);
-            }
-
-            void AssertBasePerson(BasePersonAttribute expected, BasePersonAttribute retrieved)
-            {
-                Assert.Equal(expected.Id, retrieved.Id);
-                Assert.Equal(expected.Name, retrieved.Name);
-                Assert.Equal(expected.Age, retrieved.Age);
-            }
-        }
-
-        [Fact]
-        public void TestLinqInjectTranslatesFilterUnderOuterDomain()
-        {
-            RequireServer.Check();
-
-            // Custom domain's CustomStringSerializer rewrites strings with a "test" suffix; the filter value
-            // must be serialized through the domain's registry when the Inject path is translated.
-            var customDomain = BsonSerializationDomain.CreateWithDefaultConfiguration("Test");
-            customDomain.RegisterSerializer(new CustomStringSerializer());
-
-            var client = CreateClientWithDomain(customDomain);
-            var collection = GetTypedCollection<Person>(client);
-
-            var injectedFilter = Builders<Person>.Filter.Eq(p => p.Name, "Mario");
-            var queryable = collection.AsQueryable().Where(p => injectedFilter.Inject());
-
-            var stages = Linq3TestHelpers.Translate<Person, Person>(queryable);
-            var match = stages.Single(s => s.Contains("$match"));
-            match.ToJson().Should().Contain("Mariotest");
-        }
-
-        [Fact]
-        public void TestLinqPickWithSortByRunsUnderCustomDomain()
-        {
-            RequireServer.Check();
-
-            // Smoke test: Pick's SortDefinition render path now receives the custom SerializationDomain.
-            // The assertion is that translation completes without reaching the global static registry;
-            // the resulting $bottomN reflects the expected shape.
-            var customDomain = BsonSerializationDomain.CreateWithDefaultConfiguration("Test");
-
-            var client = CreateClientWithDomain(customDomain);
-            var collection = GetTypedCollection<Person>(client);
-
-            var queryable = collection
-                .AsQueryable()
-                .GroupBy(p => p.Name)
-                .Select(g => new { Name = g.Key, YoungestAge = g.BottomN(Builders<Person>.Sort.Descending(p => p.Age), p => p.Age, 1) });
-
-            var stages = Linq3TestHelpers.Translate<Person, object>(queryable);
-            var group = stages.Single(s => s.Contains("$group"));
-            group.ToJson().Should().Contain("$bottomN");
-            group.ToJson().Should().Contain("sortBy");
-        }
-
-        [Fact]
-        public void TestLinqMathMethodsUnderCustomDomain()
-        {
-            RequireServer.Check();
-
-            // Smoke test: aggregation-expression translators that previously called BsonSerializer.LookupSerializer
-            // (Ceiling, Floor, Window aggregators, StandardDeviation, Zip) now route through context.SerializationDomain.
-            var customDomain = BsonSerializationDomain.CreateWithDefaultConfiguration("Test");
-
-            var client = CreateClientWithDomain(customDomain);
-            var collection = GetTypedCollection<NumericModel>(client);
-
-            var queryable = collection
-                .AsQueryable()
-                .Select(m => new { Ceil = Math.Ceiling(m.D), Flr = Math.Floor(m.D) });
-
-            var stages = Linq3TestHelpers.Translate<NumericModel, object>(queryable);
-            var project = stages.Single(s => s.Contains("$project"));
-            project.ToJson().Should().Contain("$ceil");
-            project.ToJson().Should().Contain("$floor");
-        }
-
-        [Fact]
-        public void TestLinqStdDevExecutableUnderCustomDomain()
-        {
-            RequireServer.Check();
-
-            // Smoke test: executable-query StandardDeviation translator now routes outputValueSerializer through the domain.
-            var customDomain = BsonSerializationDomain.CreateWithDefaultConfiguration("Test");
-
-            var client = CreateClientWithDomain(customDomain);
-            var collection = GetTypedCollection<NumericModel>(client);
-            collection.InsertMany(new[]
-            {
-                new NumericModel { Id = ObjectId.GenerateNewId(), D = 1.0 },
-                new NumericModel { Id = ObjectId.GenerateNewId(), D = 2.0 },
-                new NumericModel { Id = ObjectId.GenerateNewId(), D = 3.0 }
-            });
-
-            var stdDev = collection.AsQueryable().StandardDeviationPopulation(m => m.D);
-            stdDev.Should().BeApproximately(0.816, 0.01);
-        }
-
-        [Fact]
-        public void TestLinqOfTypeUnderCustomDomain()
-        {
-            RequireServer.Check();
-
-            // OfType in both Aggregate and AsQueryable paths uses translators that previously
-            // called BsonSerializer.LookupSerializer for the actualType. With ClassMaps registered
-            // on a custom domain, those translators must route the lookup through that domain.
-            var customDomain = BsonSerializationDomain.CreateWithDefaultConfiguration("Test");
-
-            customDomain.ClassMapRegistry.RegisterClassMap<BasePerson>(cm =>
-            {
-                cm.AutoMap();
-                cm.SetDiscriminator("bp");
-                cm.SetIsRootClass(true);
-            });
-            customDomain.ClassMapRegistry.RegisterClassMap<DerivedPerson1>(cm =>
-            {
-                cm.AutoMap();
-                cm.SetDiscriminator("dp1");
-            });
-
-            var client = CreateClientWithDomain(customDomain);
-            var collection = GetTypedCollection<BasePerson>(client);
-
-            var bp1 = new DerivedPerson1 { Name = "Alice", Age = 30, ExtraField1 = "F1" };
-            collection.InsertOne(bp1);
-
-            var ofTypeResult = collection.AsQueryable().OfType<DerivedPerson1>().Single();
-            ofTypeResult.ExtraField1.Should().Be("F1");
-        }
-
-        public class NumericModel
-        {
-            [BsonId] public ObjectId Id { get; set; }
-            public double D { get; set; }
-        }
-
-        [Fact]
-        public void TryRegisterSerializer_positive_path_under_custom_domain()
-        {
-            var customDomain = BsonSerializationDomain.CreateWithDefaultConfiguration("Test");
-            var serializer = new CustomStringSerializer("suffix");
-
-            var result = customDomain.TryRegisterSerializer(typeof(string), serializer);
-
-            result.Should().BeTrue();
-            customDomain.SerializerRegistry.GetSerializer<string>().Should().BeSameAs(serializer);
-        }
-
-        [Fact]
-        public void RegisterDiscriminatorConvention_positive_path_under_custom_domain()
-        {
-            var customDomain = BsonSerializationDomain.CreateWithDefaultConfiguration("Test");
-            var convention = new ScalarDiscriminatorConvention("_type");
-
-            customDomain.RegisterDiscriminatorConvention(typeof(BasePerson), convention);
-
-            var retrieved = customDomain.LookupDiscriminatorConvention(typeof(BasePerson));
-            retrieved.Should().BeSameAs(convention);
-        }
-
-        [Fact]
-        public void UseNullIdChecker_and_UseZeroIdChecker_are_per_domain()
-        {
-            var domain1 = BsonSerializationDomain.CreateWithDefaultConfiguration("Domain1");
-            var domain2 = BsonSerializationDomain.CreateWithDefaultConfiguration("Domain2");
-
-            domain1.UseNullIdChecker = false;
-            domain2.UseNullIdChecker = true;
-            domain1.UseZeroIdChecker = true;
-            domain2.UseZeroIdChecker = false;
-
-            domain1.UseNullIdChecker.Should().BeFalse();
-            domain2.UseNullIdChecker.Should().BeTrue();
-            domain1.UseZeroIdChecker.Should().BeTrue();
-            domain2.UseZeroIdChecker.Should().BeFalse();
-        }
-
-        [Fact]
-        public void TestMqlConstantUnderCustomDomain()
-        {
-            RequireServer.Check();
-
-            var customDomain = BsonSerializationDomain.CreateWithDefaultConfiguration("Test");
-
-            var client = CreateClientWithDomain(customDomain);
-            var collection = GetTypedCollection<EnumModel>(client);
-
-            var queryable = collection
-                .AsQueryable()
-                .Select(d => new { d.Id, V = d.ApprovalState == ApprovalState.Active ? Mql.Constant(ApprovalState.Active, BsonType.String) : ApprovalState.Inactive });
-
-            var stages = Linq3TestHelpers.Translate<EnumModel, object>(queryable);
-            var project = stages.Single(s => s.Contains("$project"));
-            project.ToJson().Should().Contain("Active");
-        }
-
-        public enum ApprovalState { Active = 1, Inactive = 2 }
-
-        public class EnumModel
-        {
-            [BsonId] public ObjectId Id { get; set; }
-            [BsonRepresentation(BsonType.String)]
-            public ApprovalState ApprovalState { get; set; }
-        }
-
-        [Fact]
-        public void Provider_created_dictionary_serializer_uses_custom_domain()
-        {
-            var customDomain = BsonSerializationDomain.CreateWithDefaultConfiguration("Test");
-            var customSerializer = new CustomStringSerializer("_X");
-            customDomain.RegisterSerializer(customSerializer);
-
-            var dictSerializer = (DictionaryInterfaceImplementerSerializer<System.Collections.Generic.Dictionary<string, string>, string, string>)
-                customDomain.SerializerRegistry.GetSerializer<System.Collections.Generic.Dictionary<string, string>>();
-
-            dictSerializer.ValueSerializer.Should().BeSameAs(customSerializer);
-            dictSerializer.KeySerializer.Should().BeSameAs(customSerializer);
-        }
-
-        [Fact]
-        public void Provider_created_readonly_dictionary_serializer_uses_custom_domain()
-        {
-            var customDomain = BsonSerializationDomain.CreateWithDefaultConfiguration("Test");
-            var customSerializer = new CustomStringSerializer("_X");
-            customDomain.RegisterSerializer(customSerializer);
-
-            var dictSerializer = (ReadOnlyDictionaryInterfaceImplementerSerializer<System.Collections.ObjectModel.ReadOnlyDictionary<string, string>, string, string>)
-                customDomain.SerializerRegistry.GetSerializer<System.Collections.ObjectModel.ReadOnlyDictionary<string, string>>();
-
-            dictSerializer.ValueSerializer.Should().BeSameAs(customSerializer);
-            dictSerializer.KeySerializer.Should().BeSameAs(customSerializer);
-        }
-
-        [Fact]
-        public void Provider_created_enumerable_serializer_uses_custom_domain()
-        {
-            var customDomain = BsonSerializationDomain.CreateWithDefaultConfiguration("Test");
-            var customSerializer = new CustomStringSerializer("_X");
-            customDomain.RegisterSerializer(customSerializer);
-
-            var listSerializer = (EnumerableInterfaceImplementerSerializer<System.Collections.Generic.List<string>, string>)
-                customDomain.SerializerRegistry.GetSerializer<System.Collections.Generic.List<string>>();
-
-            listSerializer.ItemSerializer.Should().BeSameAs(customSerializer);
-        }
-
-#if NET6_0_OR_GREATER
-        [Fact]
-        public void Provider_created_immutable_array_serializer_uses_custom_domain()
-        {
-            var customDomain = BsonSerializationDomain.CreateWithDefaultConfiguration("Test");
-            var customSerializer = new CustomStringSerializer("_X");
-            customDomain.RegisterSerializer(customSerializer);
-
-            var serializer = customDomain.SerializerRegistry.GetSerializer<System.Collections.Immutable.ImmutableArray<string>>();
-
-            serializer.Should().BeOfType<ImmutableArraySerializer<string>>();
-            ((EnumerableSerializerBase<System.Collections.Immutable.ImmutableArray<string>, string>)serializer)
-                .ItemSerializer.Should().BeSameAs(customSerializer);
-        }
-
-        [Fact]
-        public void Provider_created_immutable_dictionary_serializer_uses_custom_domain()
-        {
-            var customDomain = BsonSerializationDomain.CreateWithDefaultConfiguration("Test");
-            var customSerializer = new CustomStringSerializer("_X");
-            customDomain.RegisterSerializer(customSerializer);
-
-            var serializer = customDomain.SerializerRegistry.GetSerializer<System.Collections.Immutable.ImmutableDictionary<string, string>>();
-
-            serializer.Should().BeOfType<ImmutableDictionarySerializer<string, string>>();
-            var dictSerializer = (DictionarySerializerBase<System.Collections.Immutable.ImmutableDictionary<string, string>, string, string>)serializer;
-            dictSerializer.ValueSerializer.Should().BeSameAs(customSerializer);
-            dictSerializer.KeySerializer.Should().BeSameAs(customSerializer);
-        }
-#endif
-
-        [Fact]
-        public void Provider_created_object_serializer_uses_custom_domain()
-        {
-            var customDomain = BsonSerializationDomain.CreateWithDefaultConfiguration("Test");
-
-            var objSerializer = customDomain.SerializerRegistry.GetSerializer<object>();
-
-            objSerializer.Should().BeOfType<ObjectSerializer>();
-            ((IHasSerializationDomain)objSerializer).SerializationDomain.Should().BeSameAs(customDomain);
-        }
-
-        [Fact]
-        public void TryRegisterClassMap_with_initializer_binds_to_custom_domain()
-        {
-            var customDomain = BsonSerializationDomain.CreateWithDefaultConfiguration("Test");
-
-            var registered = customDomain.ClassMapRegistry.TryRegisterClassMap<Person>(cm => cm.AutoMap());
-
-            registered.Should().BeTrue();
-            var classMap = customDomain.ClassMapRegistry.LookupClassMap(typeof(Person));
-            classMap.SerializationDomain.Should().BeSameAs(customDomain);
-        }
-
-        private static IMongoCollection<T> GetTypedCollection<T>(IMongoClient client) =>
-            client.GetDatabase(DriverTestConfiguration.DatabaseNamespace.DatabaseName)
-                .GetCollection<T>(DriverTestConfiguration.CollectionNamespace.CollectionName);
-
-        private static IMongoCollection<BsonDocument> GetUntypedCollection(IMongoClient client) =>
-            client.GetDatabase(DriverTestConfiguration.DatabaseNamespace.DatabaseName)
-                .GetCollection<BsonDocument>(DriverTestConfiguration.CollectionNamespace.CollectionName);
-
-        private static IMongoClient CreateClientWithDomain(IBsonSerializationDomain domain, bool dropCollection = true)
-        {
-            var client = DriverTestConfiguration.CreateMongoClient(c => c.SerializationDomain = domain);
-            if (dropCollection)
-            {
-                var db = client.GetDatabase(DriverTestConfiguration.DatabaseNamespace.DatabaseName);
-                db.DropCollection(DriverTestConfiguration.CollectionNamespace.CollectionName);
-            }
-            return client;
-        }
-
-        private static IMongoClient CreateClient()
-        {
-            var client = DriverTestConfiguration.CreateMongoClient();
-            var db = client.GetDatabase(DriverTestConfiguration.DatabaseNamespace.DatabaseName);
-            db.DropCollection(DriverTestConfiguration.CollectionNamespace.CollectionName);
-            return client;
-        }
-
-        public class Person
-        {
-            [BsonId] public ObjectId Id { get; set; }
-            public string Name { get; set; }
-            public int Age { get; set; }
-        }
-
-        public class Person1
-        {
-            [BsonId] public ObjectId Id { get; set; }
-            public string Name { get; set; }
-            public int Age { get; set; }
-        }
-
-        public class BasePerson
-        {
-            [BsonId] public ObjectId Id { get; set; } = ObjectId.GenerateNewId();
-            public string Name { get; set; }
-            public int Age { get; set; }
-        }
-
-        public class DerivedPerson1 : BasePerson
-        {
-            public string ExtraField1 { get; set; }
-        }
-
-        public class DerivedPerson2 : BasePerson
-        {
-            public string ExtraField2 { get; set; }
-        }
-
-        [BsonDiscriminator("bp", RootClass = true)]
-        public class BasePersonAttribute
-        {
-            [BsonId] public ObjectId Id { get; set; } = ObjectId.GenerateNewId();
-            public string Name { get; set; }
-            public int Age { get; set; }
-        }
-
-        [BsonDiscriminator("dp1")]
-        public class DerivedPersonAttribute1 : BasePersonAttribute
-        {
-            public string ExtraField1 { get; set; }
-        }
-
-        [BsonDiscriminator("dp2")]
-        public class DerivedPersonAttribute2 : BasePersonAttribute
-        {
-            public string ExtraField2 { get; set; }
-        }
-
-        // Instantiation path: BsonSerializationProviderBase.CreateSerializer (reflection) via PrimitiveSerializationProvider
-        [Fact]
-        public void Provider_path_ObjectSerializer_uses_custom_domain()
-        {
-            var customDomain = BsonSerializationDomain.CreateWithDefaultConfiguration("Test");
-
-            var serializer = customDomain.SerializerRegistry.GetSerializer<object>();
-
-            serializer.Should().BeOfType<ObjectSerializer>();
-            ((IHasSerializationDomain)serializer).SerializationDomain.Should().BeSameAs(customDomain);
-        }
-
-        // Instantiation path: BsonSerializationProviderBase.CreateSerializer (reflection) via PrimitiveSerializationProvider
-        [Fact]
-        public void Provider_path_TupleSerializer_resolves_items_from_custom_domain()
-        {
-            var customDomain = BsonSerializationDomain.CreateWithDefaultConfiguration("Test");
-            var customSerializer = new CustomStringSerializer("_X");
-            customDomain.RegisterSerializer(customSerializer);
-
-            var serializer = customDomain.SerializerRegistry.GetSerializer<Tuple<string>>();
-
-            serializer.Should().BeOfType<TupleSerializer<string>>();
-            ((TupleSerializer<string>)serializer).Item1Serializer.Should().BeSameAs(customSerializer);
-        }
-
-        // Instantiation path: BsonSerializationProviderBase.CreateSerializer (reflection) via PrimitiveSerializationProvider
-        [Fact]
-        public void Provider_path_ValueTupleSerializer_resolves_items_from_custom_domain()
-        {
-            var customDomain = BsonSerializationDomain.CreateWithDefaultConfiguration("Test");
-            var customSerializer = new CustomStringSerializer("_X");
-            customDomain.RegisterSerializer(customSerializer);
-
-            var serializer = customDomain.SerializerRegistry.GetSerializer<ValueTuple<string>>();
-
-            serializer.Should().BeOfType<ValueTupleSerializer<string>>();
-            ((ValueTupleSerializer<string>)serializer).Item1Serializer.Should().BeSameAs(customSerializer);
-        }
-
-        // Instantiation path: BsonSerializationProviderBase.CreateSerializer (reflection) via CollectionsSerializationProvider
-        [Fact]
-        public void Provider_path_ExpandoObjectSerializer_uses_custom_domain()
-        {
-            var customDomain = BsonSerializationDomain.CreateWithDefaultConfiguration("Test");
-
-            var serializer = customDomain.SerializerRegistry.GetSerializer<System.Dynamic.ExpandoObject>();
-
-            serializer.Should().BeOfType<ExpandoObjectSerializer>();
-            ((IHasSerializationDomain)serializer).SerializationDomain.Should().BeSameAs(customDomain);
-        }
-
-        // Instantiation path: DiscriminatedInterfaceSerializationProvider via BsonSerializationProviderBase.CreateSerializer (reflection)
-        [Fact]
-        public void Provider_path_DiscriminatedInterfaceSerializer_uses_custom_domain()
-        {
-            var customDomain = BsonSerializationDomain.CreateWithDefaultConfiguration("Test");
-
-            var serializer = customDomain.SerializerRegistry.GetSerializer<IAnimal>();
-
-            serializer.Should().BeOfType<DiscriminatedInterfaceSerializer<IAnimal>>();
-            ((IHasSerializationDomain)serializer).SerializationDomain.Should().BeSameAs(customDomain);
-        }
-
-        // Instantiation path: DiscriminatedInterfaceSerializer internally chains
-        // ObjectSerializer.WithDiscriminatorConvention().WithAllowedTypes() — verifies the With* fix
-        [Fact]
-        public void Provider_path_DiscriminatedInterfaceSerializer_internal_ObjectSerializer_preserves_domain()
-        {
-            var customDomain = BsonSerializationDomain.CreateWithDefaultConfiguration("Test");
-
-            var serializer = (DiscriminatedInterfaceSerializer<IAnimal>)
-                customDomain.SerializerRegistry.GetSerializer<IAnimal>();
-
-            var objectSerializerField = typeof(DiscriminatedInterfaceSerializer<IAnimal>)
-                .GetField("_objectSerializer", System.Reflection.BindingFlags.NonPublic | System.Reflection.BindingFlags.Instance);
-            var objectSerializer = (ObjectSerializer)objectSerializerField.GetValue(serializer);
-
-            ((IHasSerializationDomain)objectSerializer).SerializationDomain.Should().BeSameAs(customDomain);
-        }
-
-        // Instantiation path: CollectionsSerializationProvider via BsonSerializationProviderBase.CreateSerializer (reflection)
-        [Fact]
-        public void Provider_path_ImpliedImplementationInterfaceSerializer_uses_custom_domain()
-        {
-            var customDomain = BsonSerializationDomain.CreateWithDefaultConfiguration("Test");
-
-            var serializer = customDomain.SerializerRegistry.GetSerializer<System.Collections.Generic.IDictionary<string, string>>();
-
-            serializer.Should().BeOfType<ImpliedImplementationInterfaceSerializer<System.Collections.Generic.IDictionary<string, string>, System.Collections.Generic.Dictionary<string, string>>>();
-            ((IHasSerializationDomain)serializer).SerializationDomain.Should().BeSameAs(customDomain);
-        }
-
-        // Instantiation path: BsonClassMapSerializationProvider (Activator.CreateInstance with domain)
-        [Fact]
-        public void ClassMap_provider_path_uses_custom_domain()
-        {
-            var customDomain = BsonSerializationDomain.CreateWithDefaultConfiguration("Test");
-
-            var serializer = customDomain.SerializerRegistry.GetSerializer<Person>();
-
-            serializer.Should().BeOfType<BsonClassMapSerializer<Person>>();
-            ((IHasSerializationDomain)serializer).SerializationDomain.Should().BeSameAs(customDomain);
-        }
-
-        // Instantiation path: AttributedSerializationProvider → BsonSerializerAttribute.CreateSerializer
-        [Fact]
-        public void Attribute_path_parameterless_serializer_resolves_from_custom_domain()
-        {
-            var customDomain = BsonSerializationDomain.CreateWithDefaultConfiguration("Test");
-
-            var serializer = customDomain.SerializerRegistry.GetSerializer<TypeWithAttributedSerializer>();
-
-            serializer.Should().BeOfType<CustomAttributeSerializer>();
-        }
-
-        // Instantiation path: AttributedSerializationProvider → BsonSerializerAttribute.CreateSerializer with domain ctor
-        [Fact]
-        public void Attribute_path_domain_ctor_serializer_uses_custom_domain()
-        {
-            var customDomain = BsonSerializationDomain.CreateWithDefaultConfiguration("Test");
-
-            var serializer = customDomain.SerializerRegistry.GetSerializer<TypeWithDomainAwareAttributedSerializer>();
-
-            serializer.Should().BeOfType<DomainAwareAttributeSerializer>();
-            ((IHasSerializationDomain)serializer).SerializationDomain.Should().BeSameAs(customDomain);
-        }
-
-        // With* path: ObjectSerializer.WithDiscriminatorConvention preserves domain
-        [Fact]
-        public void WithDiscriminatorConvention_preserves_domain()
-        {
-            var customDomain = BsonSerializationDomain.CreateWithDefaultConfiguration("Test");
-            var objectSerializer = (ObjectSerializer)customDomain.SerializerRegistry.GetSerializer<object>();
-
-            var reconfigured = objectSerializer.WithDiscriminatorConvention(new ScalarDiscriminatorConvention("_t"));
-
-            ((IHasSerializationDomain)reconfigured).SerializationDomain.Should().BeSameAs(customDomain);
-        }
-
-        // With* path: ObjectSerializer.WithAllowedTypes preserves domain
-        [Fact]
-        public void WithAllowedTypes_preserves_domain()
-        {
-            var customDomain = BsonSerializationDomain.CreateWithDefaultConfiguration("Test");
-            var objectSerializer = (ObjectSerializer)customDomain.SerializerRegistry.GetSerializer<object>();
-
-            var reconfigured = objectSerializer.WithAllowedTypes(t => true, t => true);
-
-            ((IHasSerializationDomain)reconfigured).SerializationDomain.Should().BeSameAs(customDomain);
-        }
-
-        // With* path: EnumerableInterfaceImplementerSerializer.WithItemSerializer preserves domain
-        [Fact]
-        public void WithItemSerializer_on_enumerable_preserves_domain()
-        {
-            var customDomain = BsonSerializationDomain.CreateWithDefaultConfiguration("Test");
-            var listSerializer = (EnumerableInterfaceImplementerSerializer<System.Collections.Generic.List<string>, string>)
-                customDomain.SerializerRegistry.GetSerializer<System.Collections.Generic.List<string>>();
-
-            var domainField = typeof(EnumerableSerializerBase<System.Collections.Generic.List<string>, string>)
-                .GetField("_serializationDomain", System.Reflection.BindingFlags.NonPublic | System.Reflection.BindingFlags.Instance);
-
-            domainField.GetValue(listSerializer).Should().BeSameAs(customDomain);
-
-            var reconfigured = listSerializer.WithItemSerializer(new CustomStringSerializer("_Y"));
-            domainField.GetValue(reconfigured).Should().BeSameAs(customDomain);
-        }
-
-        // With* path: ImpliedImplementationInterfaceSerializer.WithImplementationSerializer preserves domain
-        [Fact]
-        public void WithImplementationSerializer_preserves_domain()
-        {
-            var customDomain = BsonSerializationDomain.CreateWithDefaultConfiguration("Test");
-            var serializer = (ImpliedImplementationInterfaceSerializer<System.Collections.Generic.IDictionary<string, string>, System.Collections.Generic.Dictionary<string, string>>)
-                customDomain.SerializerRegistry.GetSerializer<System.Collections.Generic.IDictionary<string, string>>();
-
-            ((IHasSerializationDomain)serializer).SerializationDomain.Should().BeSameAs(customDomain);
-
-            var newImplSerializer = (IBsonSerializer<System.Collections.Generic.Dictionary<string, string>>)
-                customDomain.SerializerRegistry.GetSerializer<System.Collections.Generic.Dictionary<string, string>>();
-            var reconfigured = serializer.WithImplementationSerializer(newImplSerializer);
-
-            ((IHasSerializationDomain)reconfigured).SerializationDomain.Should().BeSameAs(customDomain);
-        }
-
-        // Reconfiguration path: SerializerConfigurator.ReconfigureSerializerRecursively via ObjectSerializerAllowedTypesConvention
-        [Fact]
-        public void ObjectSerializerAllowedTypesConvention_preserves_domain_on_reconfigured_ObjectSerializer()
-        {
-            var customDomain = BsonSerializationDomain.CreateWithDefaultConfiguration("Test");
-
-            var pack = new ConventionPack(customDomain);
-            pack.Add(ObjectSerializerAllowedTypesConvention.AllowAllTypes);
-            customDomain.ConventionRegistry.Register("allow-all", pack, t => t == typeof(ClassWithObjectMember));
-
-            // Force the class map to be created and frozen so the convention is applied
-            customDomain.SerializerRegistry.GetSerializer<ClassWithObjectMember>();
-
-            var classMap = customDomain.ClassMapRegistry.LookupClassMap(typeof(ClassWithObjectMember));
-            var dataMemberMap = classMap.GetMemberMap("Data");
-            var memberSerializer = dataMemberMap.GetSerializer();
-
-            memberSerializer.Should().BeOfType<ObjectSerializer>();
-            ((IHasSerializationDomain)memberSerializer).SerializationDomain.Should().BeSameAs(customDomain);
-        }
-
-        public interface IAnimal
-        {
-            string Name { get; set; }
-        }
-
-        public class Dog : IAnimal
-        {
-            public string Name { get; set; }
-            public string Breed { get; set; }
-        }
-
-        [BsonSerializer(typeof(CustomAttributeSerializer))]
-        public class TypeWithAttributedSerializer
-        {
-            public string Value { get; set; }
-        }
-
-        public class CustomAttributeSerializer : SerializerBase<TypeWithAttributedSerializer>
-        {
-            public override TypeWithAttributedSerializer Deserialize(BsonDeserializationContext context, BsonDeserializationArgs args)
-            {
-                context.Reader.ReadStartDocument();
-                context.Reader.ReadString();
-                context.Reader.ReadEndDocument();
-                return new TypeWithAttributedSerializer { Value = "deserialized" };
-            }
-
-            public override void Serialize(BsonSerializationContext context, BsonSerializationArgs args, TypeWithAttributedSerializer value)
-            {
-                context.Writer.WriteStartDocument();
-                context.Writer.WriteName("v");
-                context.Writer.WriteString(value.Value ?? "");
-                context.Writer.WriteEndDocument();
-            }
-        }
-
-        [BsonSerializer(typeof(DomainAwareAttributeSerializer))]
-        public class TypeWithDomainAwareAttributedSerializer
-        {
-            public string Value { get; set; }
-        }
-
-        public class DomainAwareAttributeSerializer : SerializerBase<TypeWithDomainAwareAttributedSerializer>, IHasSerializationDomain
-        {
-            private readonly IBsonSerializationDomain _serializationDomain;
-
-            public DomainAwareAttributeSerializer()
-                : this(BsonSerializationDomain.Default)
-            {
-            }
-
-            internal DomainAwareAttributeSerializer(IBsonSerializationDomain serializationDomain)
-            {
-                _serializationDomain = serializationDomain;
-            }
-
-            IBsonSerializationDomain IHasSerializationDomain.SerializationDomain => _serializationDomain;
-
-            public override TypeWithDomainAwareAttributedSerializer Deserialize(BsonDeserializationContext context, BsonDeserializationArgs args)
-            {
-                context.Reader.ReadStartDocument();
-                context.Reader.ReadString();
-                context.Reader.ReadEndDocument();
-                return new TypeWithDomainAwareAttributedSerializer { Value = "deserialized" };
-            }
-
-            public override void Serialize(BsonSerializationContext context, BsonSerializationArgs args, TypeWithDomainAwareAttributedSerializer value)
-            {
-                context.Writer.WriteStartDocument();
-                context.Writer.WriteName("v");
-                context.Writer.WriteString(value.Value ?? "");
-                context.Writer.WriteEndDocument();
-            }
-        }
-
-        public class ClassWithObjectMember
-        {
-            public ObjectId Id { get; set; }
-            public object Data { get; set; }
-        }
-
-        // This serializer adds the _appended variable to any serialised string
-        public class CustomStringSerializer(string appended = "test")
-            : SealedClassSerializerBase<string>
-        {
-            /// <inheritdoc/>
-            public override int GetHashCode() => 0;
-
-            protected override string DeserializeValue(BsonDeserializationContext context, BsonDeserializationArgs args)
-            {
-                var bsonReader = context.Reader;
-
-                var bsonType = bsonReader.GetCurrentBsonType();
-                return bsonType switch
-                {
-                    BsonType.String => bsonReader.ReadString().Replace(appended, ""),
-                    _ => throw CreateCannotDeserializeFromBsonTypeException(bsonType)
-                };
-            }
-
-            protected override void SerializeValue(BsonSerializationContext context, BsonSerializationArgs args, string value)
-            {
-                var bsonWriter = context.Writer;
-                bsonWriter.WriteString(value + appended);
-            }
-        }
-
-        public class CustomObjectIdGenerator : IIdGenerator
-        {
-            public object GenerateId(object container, object document)
-            {
-                return  ObjectId.Parse("6797b56bf5495bf53aa3078f");
-            }
-
-            public bool IsEmpty(object id)
-            {
-                return true;
-            }
-        }
-    }
-}
